(cleanup) breaking out compressor from conn
